Refactor : STT 디렉토리 생성
STT 관련 파일을 분리하기 위해 Cam/STT 디렉토리를 생성하였습니다.
